Remove config, add flag
