+++ src/powerbox/tools.py
@@ -9,7 +9,6 @@ import warnings
 
 from . import dft
 
-
 def _getbins(bins, coords, log):
     mx = coords.max()
     if not np.iterable(bins):
@@ -31,6 +30,7 @@ def angular_average(
     bin_ave=True,
     get_variance=False,
     log_bins=False,
+    threads=None,
 ):
     r"""
     Average a given field within radial bins.
@@ -383,6 +383,7 @@ def get_power(
     log_bins=False,
     ignore_zero_mode=False,
     k_weights=1,
+    threads=None,
 ):
     r"""
     Calculate isotropic power spectrum of a field, or cross-power of two similar fields.
@@ -537,9 +538,9 @@ def get_power(
     V = np.prod(boxlength)
 
     # Calculate the n-D power spectrum and align it with the k from powerbox.
-    FT, freq, k = dft.fft(deltax, L=boxlength, a=a, b=b, ret_cubegrid=True)
+    FT, freq, k = dft.fft(deltax, L=boxlength, a=a, b=b, ret_cubegrid=True, threads=threads)
 
-    FT2 = dft.fft(deltax2, L=boxlength, a=a, b=b)[0] if deltax2 is not None else FT
+    FT2 = dft.fft(deltax2, L=boxlength, a=a, b=b,threads=threads)[0] if deltax2 is not None else FT
     P = np.real(FT * np.conj(FT2) / V**2)
 
     if vol_normalised_power:
